refactor(validator): input을 담당하는 클래스에서 validate 부분을 클래스로 분리
