+++ src/main/java/vendingmachine/view/input/MoneyInputView.java
@@ -1,7 +1,7 @@
 package vendingmachine.view.input;
 
 import camp.nextstep.edu.missionutils.Console;
-import vendingmachine.constant.Condition;
+import vendingmachine.validator.MoneyValidator;
 import vendingmachine.constant.Input;
 
 public class MoneyInputView {
@@ -11,7 +11,8 @@ public class MoneyInputView {
             print(Input.COIN_MONEY_GUIDE_MESSAGE.getText());
             String money = Console.readLine();
             try {
-                tryToInputMoneyForMakeCoin(money);
+                MoneyValidator moneyValidator = new MoneyValidator();
+                moneyValidator.tryToInputMoneyForMakeCoin(money);
                 return Integer.parseInt(money);
             } catch (IllegalArgumentException e) {
             }
@@ -24,52 +25,11 @@ public class MoneyInputView {
             print(Input.PURCHASE_MONEY_GUIDE_MESSAGE.getText());
             String money = Console.readLine();
             try {
-                tryInputMoneyForPurchase(money);
+                MoneyValidator moneyValidator = new MoneyValidator();
+                moneyValidator.tryInputMoneyForPurchase(money);
                 return Integer.parseInt(money);
             } catch (IllegalArgumentException e) {
             }
-        }
-    }
-
-    private void tryToInputMoneyForMakeCoin(String money) throws IllegalArgumentException {
-        validateCoinMoneyDigit(money);
-        validateCoinMoneyBlank(money);
-    }
-
-    private void tryInputMoneyForPurchase(String money) throws IllegalArgumentException {
-        validatePurchaseMoneyDigit(money);
-        validatePurchaseMoneyBlank(money);
-    }
-
-    private void validateCoinMoneyDigit(String money) {
-        for (int m = 0; m < money.length(); m++) {
-            if (!Character.isDigit(money.charAt(m))) {
-                print(Input.COIN_MONEY_DIGIT_ERROR_MESSAGE.getText());
-                throw new IllegalArgumentException();
-            }
-        }
-    }
-
-    private void validateCoinMoneyBlank(String money) {
-        if (money.length() == Condition.LENGTH_0.getNumber()) {
-            print(Input.COIN_MONEY_LENGTH_0_ERROR_MESSAGE.getText());
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validatePurchaseMoneyDigit(String money) {
-        for (int m = 0; m < money.length(); m++) {
-            if (!Character.isDigit(money.charAt(m))) {
-                print(Input.PURCHASE_MONEY_DIGIT_ERROR_MESSAGE.getText());
-                throw new IllegalArgumentException();
-            }
-        }
-    }
-
-    private void validatePurchaseMoneyBlank(String money) {
-        if (money.length() == Condition.LENGTH_0.getNumber()) {
-            print(Input.PURCHASE_MONEY_LENGTH_0_ERROR_MESSAGE.getText());
-            throw new IllegalArgumentException();
         }
     }
