+++ src/main/java/vendingmachine/view/input/PurchaseInputView.java
@@ -1,14 +1,10 @@
 package vendingmachine.view.input;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 import camp.nextstep.edu.missionutils.Console;
-import vendingmachine.constant.Condition;
+import vendingmachine.validator.PurchaseValidator;
 import vendingmachine.constant.Input;
 import vendingmachine.constant.Output;
 import vendingmachine.domain.Money;
-import vendingmachine.domain.Product;
 import vendingmachine.domain.ProductRepository;
 import vendingmachine.domain.ReturnCoin;
 
@@ -29,51 +25,17 @@ public class PurchaseInputView {
 
             String productName = inputInit();
             try {
-                tryToInputProductForPurchase(productName);
-                ProductRepository.getInstance().purchaseProduct(productName);
-                Money.getInstance().minusMoney(productName);
+                PurchaseValidator purchaseValidator = new PurchaseValidator();
+                purchaseValidator.tryToInputProductForPurchase(productName);
+                updateResourceState(productName);
             } catch (IllegalArgumentException e) {
             }
         }
     }
 
-    private void tryToInputProductForPurchase(String productName) throws IllegalArgumentException{
-        validateProductExist(productName);
-        validateProductAmountExist(productName);
-        validateEnoughMoneyForProductCost(productName);
-    }
-
-    private void validateProductAmountExist(String productName) {
-        List<Product> products = ProductRepository.getInstance().getProducts();
-        List<Product> findProduct = products.stream().
-                filter(product -> product.getName().equals(productName)).collect(Collectors.toList());
-
-        if (findProduct.get(Condition.INDEX_0.getNumber()).getQuantity() == 0) {
-            System.out.println(Input.PRODUCT_AMOUNT_LACK_ERROR_MESSAGE.getText());
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validateProductExist(String productName) {
-        List<Product> products = ProductRepository.getInstance().getProducts();
-        boolean result = products.stream()
-                .anyMatch(product -> product.getName().equals(productName));
-
-        if (!result) {
-            System.out.println(Input.PRODUCT_EXIST_ERROR_MESSAGE.getText());
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validateEnoughMoneyForProductCost(String productName) {
-        List<Product> products = ProductRepository.getInstance().getProducts();
-        List<Product> findProduct = products.stream().
-                filter(product -> product.getName().equals(productName)).collect(Collectors.toList());
-
-        if (findProduct.get(Condition.INDEX_0.getNumber()).getCost() > Money.getInstance().getMoney()) {
-            System.out.println(Input.PRODUCT_PURCHASE_LACK_MONEY_ERROR_MESSAGE.getText());
-            throw new IllegalArgumentException();
-        }
+    private void updateResourceState(String productName) {
+        ProductRepository.getInstance().purchaseProduct(productName);
+        Money.getInstance().minusMoney(productName);
     }
 
     private void divisionLine() {
